Refactor type factory files and methods
