+++ src/types/factories/AbstractTypeFactory.ts
@@ -1,0 +1,18 @@
+export abstract class AbstractTypeFactory {
+
+  constructor() {
+  }
+
+  static factory() {
+    throw new Error("Method Not Implemented! Use constructor.");
+  }
+
+  create() {
+    throw new Error("Method Not Implemented! Use constructor.");
+  }
+
+
+}
+
+export class TestFactory extends AbstractTypeFactory {};
+
