+++ src/types/factories/README.md
@@ -1,0 +1,36 @@
+# Type Factories
+
+## Factories overview
+Factories in this project make creating type objects more easily.
+
+Each factory class offers a static `factory` method that allows for the creation of new factory instances. Additionally,
+an instance's `create` method generates an object of the type that factory produces.
+
+Each factory contains its own additional functions based on the type it produces.
+
+## Contents
+
+- [AbstractTypeFactory](#abstracttypefactory)
+- [MessageFactories](#messagefactories)
+  - [JetstreamMessageFactory](#jetstreammessagefactory)
+  - [CreateMessageFactory](#createmessagefactory)
+  - [CreateSkeetMessageFactory](#createskeetmessagefactory)
+- [RecordFactories](#recordfactories)
+  - [CreateSkeetRecordFactory](#createskeetrecordfactory)
+  - [ReplyFactory](#replyfactory)
+
+## AbstractTypeFactory
+
+## MessageFactories
+
+### JetstreamMessageFactory
+
+### CreateMessageFactory
+
+### CreateSkeetMessageFactory
+
+## RecordFactories
+
+### CreateSkeetRecordFactory
+
+### ReplyFactory
